backend/pyproject.toml
[tool.poetry]
name = "funnyalexander"
version = "0.1.0"
description = ""
authors = ["Nikita Semenov <nasemenov726@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
alembic = "^1.13.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.3.3"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
